requirements.txt
Flask
Flask-Caching
gunicorn
google-cloud-firestore
google-cloud-storage
google-cloud-texttospeech
google-cloud-tasks
trafilatura
newspaper3k
beautifulsoup4
readability-lxml
python-dotenv
Flask-Talisman
pydub
python-json-logger
python-dateutil
cachetools
feedgen
feedparser

requests
